[Tools:Box:Search] Adding agent examples
