+++ llama-index-integrations/tools/llama-index-tools-box/tests/test_tools_box_search.py
@@ -1,6 +1,9 @@
 from box_sdk_gen import BoxClient
+import openai
+import pytest
 
-from llama_index.tools.box import BoxSearchToolSpec, BoxSearchByMetadataToolSpec
+from llama_index.agent.openai import OpenAIAgent
+from llama_index.tools.box import BoxSearchToolSpec
 
 from tests.conftest import get_testing_data
 
@@ -8,31 +11,26 @@ from tests.conftest import get_testing_data
 def test_box_tool_search(box_client_ccg_integration_testing: BoxClient):
     query = "invoice"
     box_tool = BoxSearchToolSpec(box_client_ccg_integration_testing)
-    docs = box_tool.search(query=query)
+    docs = box_tool.box_search(query=query)
     assert len(docs) > 0
 
 
-def test_box_tool_search_by_metadata(box_client_ccg_integration_testing: BoxClient):
+def test_box_tool_search_agent(box_client_ccg_integration_testing: BoxClient):
     test_data = get_testing_data()
+    openai_api_key = test_data["openai_api_key"]
 
-    # Parameters
-    from_ = (
-        test_data["metadata_enterprise_scope"]
-        + "."
-        + test_data["metadata_template_key"]
-    )
-    ancestor_folder_id = test_data["test_folder_invoice_po_id"]
-    query = "documentType = :docType "
-    query_params = {"docType": "Invoice"}
+    if openai_api_key is None:
+        raise pytest.skip("OpenAI API key is not provided.")
 
-    box_tool = BoxSearchByMetadataToolSpec(
-        box_client=box_client_ccg_integration_testing
+    box_tool_spec = BoxSearchToolSpec(box_client_ccg_integration_testing)
+    tool = box_tool_spec.to_tool_list()[0]
+
+    openai.api_key = openai_api_key
+
+    agent = OpenAIAgent.from_tools(
+        box_tool_spec.to_tool_list(),
+        verbose=True,
     )
 
-    docs = box_tool.search(
-        from_=from_,
-        ancestor_folder_id=ancestor_folder_id,
-        query=query,
-        query_params=query_params,
-    )
-    assert len(docs) > 0
+    answer = agent.chat("search all invoices")
+    assert answer is not None
